feat(Quantidade): alterar quantidade para generalizar os tipos de produtos.

+++ modulo-02/exercicio01.sql
@@ -15,7 +15,7 @@ CREATE TABLE produto (
 	nome_descritivo VARCHAR(100) NOT NULL,
 	data_criacao DATE NOT NULL,
 	local_no_estoque VARCHAR(30),
-	quantidade INT NOT NULL,
+	quantidade DECIMAL (7,3) NOT NULL,
 	preco FLOAT NOT NULL,
 	CONSTRAINT PK_Produto PRIMARY KEY (id_produto)
 );
@@ -26,5 +26,3 @@ VALUES (1, 'Computador', 'Computador Dell', '28/04/17', 'A12', 3, 2800);
 
 INSERT INTO produto (id_produto, nome_curto, nome_descritivo, data_criacao, local_no_estoque, quantidade, preco)
 VALUES (2, 'Celular', 'Zenfone 2', '28/04/17', 'B20', 5, 1000);
-
-
